mavros #302: split UAS impl by function blocks
