+++ mavros/CMakeLists.txt
@@ -154,11 +154,13 @@ include_directories(
 
 
 add_library(mavros
-  src/lib/uas.cpp
+  src/lib/uas_data.cpp
+  src/lib/uas_stringify.cpp
+  src/lib/uas_timesync.cpp
+  src/lib/uas_sensor_orientation.cpp
   src/lib/mavros.cpp
   src/lib/mavlink_diag.cpp
   src/lib/rosconsole_bridge.cpp
-  src/lib/sensor_orientation.cpp
 )
 add_dependencies(mavros
   mavros_generate_messages_cpp
